Address review: RAII quant-flag guard in tests, log underflow, undo apply cleanup
The q16 testing flag is now reset by a scope guard, so a failed
assertion or a skip no longer leaks the flag into later tests. The
quantize log casts before subtracting so a size increase reports
negative instead of underflowing. ShNCanonicalRowsUndoEntry::apply
checks the scene once up front and drops the dead trailing guard.

+++ src/visualizer/scene/viewer_splat_quantize.cpp
@@ -63,7 +63,8 @@ namespace lfs::vis {
         LOG_INFO(
             "Viewer PLY SH q16: path='{}' gaussians={} before_bytes={} after_bytes={} saved_mib={:.3f}",
             lfs::core::path_to_utf8(path), model.size(), shN_before_bytes, shN_after_bytes,
-            static_cast<double>(shN_before_bytes - shN_after_bytes) / (1024.0 * 1024.0));
+            (static_cast<double>(shN_before_bytes) - static_cast<double>(shN_after_bytes)) /
+                (1024.0 * 1024.0));
     }
 
 } // namespace lfs::vis
